run_mission example to work with malmo package
