+++ Malmo/samples/Python_examples/malmoutils.py
@@ -18,7 +18,13 @@
 # ------------------------------------------------------------------------------------------------
 
 from __future__ import print_function
-import MalmoPython
+# Allow MalmoPython to be imported both from an installed 
+# malmo module and separately as a native library.
+try:
+    import malmo.MalmoPython as MalmoPython
+except ImportError:
+    import MalmoPython
+
 import os
 import sys
 import errno
@@ -37,13 +43,15 @@ def fix_print():
         import builtins
         builtins.print = functools.partial(print, flush=True)
 
-def parse_command_line(agent_host):
+def parse_command_line(agent_host, argv=None):
+    if argv is None:
+       argv = sys.argv
     # Add standard options required by test suite:
     agent_host.addOptionalStringArgument( "recording_dir,r", "Path to location for saving mission recordings", "" )
     agent_host.addOptionalFlag( "record_video,v", "Record video stream" )
     # Attempt to parse:
     try:
-        agent_host.parse( sys.argv )
+        agent_host.parse(argv)
     except RuntimeError as e:
         print('ERROR:',e)
         print(agent_host.getUsage())
@@ -91,3 +99,4 @@ def get_recordings_directory(agent_host):
             if exception.errno != errno.EEXIST: # ignore error if already existed
                 raise
     return recordingsDirectory
+
