+++ scripts/docker/manylinux.dockerfile
@@ -18,16 +18,18 @@
 # NB if building this on Windows/OSX, make sure Docker has been allowed enough memory - the default 2048Mb is not
 # enough for the gradle Minecraft deobfuscation step.
 
-# https://github.com/pypa/manylinux
-# Not everything is in place. Use the following to build:
+# This dockerfile supports building a "manylinux1" (hopefully somewhat widely compatible) MalmoPython.so to use
+# with Linux flavours  (please see https://github.com/pypa/manylinux for detials).
+# Not everything is in place as Centos5 is past end of life. Use the following to build after launching a container interactively:
 # cmake -DSTATIC_BOOST=ON -DBoost_INCLUDE_DIR=/home/malmo/boost/boost_1_66_0/include -DUSE_PYTHON_VERSIONS=3.5 -DCMAKE_BUILD_TYPE=Release -DBUILD_DOCUMENTATION=OFF -DINCLUDE_JAVA=OFF -DBUILD_MOD=OFF ..
 
-FROM quay.io/pypa/manylinux1_x86_64 
+FROM quay.io/pypa/manylinux1_x86_64
 
 # Need to install sudo first:
 RUN yum install -y sudo
 # Can only run sudo from a terminal on CentOS 5.
 
+# (Skip) desktop components.
 #RUN sudo rpm --import http://li.nux.ro/download/nux/RPM-GPG-KEY-nux.ro
 #RUN sudo rpm -Uvh http://li.nux.ro/download/nux/dextop/el7/x86_64/nux-dextop-release-0-5.el7.nux.noarch.rpm
 
@@ -38,7 +40,7 @@ RUN bash -c 'echo "malmo ALL=(ALL:ALL) NOPASSWD: ALL" | (EDITOR="tee -a" visudo)
 # Centos 5 is end of life. Yum retired to:
 COPY ./centos5.yum.repos.d/ /etc/yum.repos.d/
 
-# Add repo for xpra:
+# (Skip) Add repo for xpra:
 #RUN rpm --import "http://winswitch.org/gpg.asc"
 #RUN su -c 'curl https://winswitch.org/downloads/CentOS/winswitch.repo | tee /etc/yum.repos.d/winswitch.repo'
 
@@ -81,6 +83,7 @@ WORKDIR /home/malmo/Python-3.5.5
 RUN ./configure --enable-shared
 RUN make
 
+# Switch to root as sudo complains here on CentOS 5.
 USER 0
 WORKDIR /home/malmo/Python-3.5.5
 RUN make install
@@ -91,7 +94,8 @@ USER malmo
 RUN mkdir /home/malmo/boost
 WORKDIR /home/malmo/boost
 RUN echo "using python : 3.5 : /usr/local/bin/python3 : /usr/local/include/python3.5m : /usr/local/lib ;" > ~/user-config.jam
-# wget is failing with "moved permanently"
+# wget is failing with "moved permanently". Copy in tar file instead.
+# wget http://sourceforge.net/projects/boost/files/boost/1.66.0/boost_1_66_0.tar.gz
 COPY boost_1_66_0.tar.gz .
 RUN tar xvf boost_1_66_0.tar.gz
 WORKDIR /home/malmo/boost/boost_1_66_0
@@ -110,7 +114,8 @@ RUN make -j4
 USER 0
 RUN make install
 
-#RUN pip3 install future && sudo pip3 install pillow && sudo pip3 install matplotlib
+# (Skip as no pip3)
+# RUN pip3 install future && sudo pip3 install pillow && sudo pip3 install matplotlib
 
 RUN yum update -y && yum -y install dos2unix
 COPY ./build.sh /home/malmo
